+++ frontend/graphql/queries/courses.ts
@@ -148,3 +148,75 @@ export const HandlerCoursesQuery = gql`
     }
   }
 `
+
+export const CourseQuery = gql`
+  query CourseDetails($slug: String) {
+    course(slug: $slug) {
+      id
+      name
+      slug
+      ects
+      order
+      study_module_order
+      teacher_in_charge_name
+      teacher_in_charge_email
+      support_email
+      start_date
+      end_date
+      tier
+      photo {
+        id
+        compressed
+        compressed_mimetype
+        uncompressed
+        uncompressed_mimetype
+      }
+      promote
+      start_point
+      hidden
+      study_module_start_point
+      status
+      course_translations {
+        id
+        name
+        language
+        description
+        link
+      }
+      open_university_registration_links {
+        id
+        course_code
+        language
+        link
+      }
+      study_modules {
+        id
+      }
+      course_variants {
+        id
+        slug
+        description
+      }
+      course_aliases {
+        id
+        course_code
+      }
+      inherit_settings_from {
+        id
+      }
+      completions_handled_by {
+        id
+      }
+      has_certificate
+      user_course_settings_visibilities {
+        id
+        language
+      }
+      upcoming_active_link
+      automatic_completions
+      automatic_completions_eligible_for_ects
+      exercise_completions_needed
+      points_needed
+    }
+  }
+`
