+++ frontend/pages/[lng]/courses/new.tsx
@@ -1,6 +1,5 @@
 import { useEffect, useState } from "react"
 import { WideContainer } from "/components/Container"
-import { useQuery } from "@apollo/client"
 import CourseEdit from "/components/Dashboard/Editor/Course"
 import CourseEdit2 from "/components/Dashboard/Editor2/Course"
 import FormSkeleton from "/components/Dashboard/Editor/FormSkeleton"
@@ -8,20 +7,11 @@ import { H1NoBackground } from "/components/Text/headers"
 import ModifiableErrorMessage from "/components/ModifiableErrorMessage"
 import withAdmin from "/lib/with-admin"
 import CoursesTranslations from "/translations/courses"
-import { CourseEditorStudyModules } from "/static/types/generated/CourseEditorStudyModules"
-import {
-  CourseEditorStudyModuleQuery,
-  CourseEditorCoursesQuery,
-} from "/graphql/queries/courses"
-import { CourseEditorCourses } from "/static/types/generated/CourseEditorCourses"
 import notEmpty from "/util/notEmpty"
 import { useQueryParameter } from "/util/useQueryParameter"
-import { CourseQuery } from "/pages/[lng]/courses/[slug]/edit"
-import {
-  CourseDetails,
-  CourseDetails_course,
-} from "/static/types/generated/CourseDetails"
+import { CourseDetails_course } from "/static/types/generated/CourseDetails"
 import { useTranslator } from "/util/useTranslator"
+import { useEditorCourses } from "/hooks/useEditorCourses"
 
 function stripId<T>(data: T): T {
   if (data === null || data === undefined) return data
@@ -48,22 +38,13 @@ const NewCourse = () => {
   const beta = useQueryParameter("beta", false)
 
   const {
-    data: studyModulesData,
-    loading: studyModulesLoading,
-    error: studyModulesError,
-  } = useQuery<CourseEditorStudyModules>(CourseEditorStudyModuleQuery)
-  const {
-    data: coursesData,
-    loading: coursesLoading,
-    error: coursesError,
-  } = useQuery<CourseEditorCourses>(CourseEditorCoursesQuery)
-  const {
-    data: courseData,
-    loading: courseLoading,
-    // @ts-ignore: for now
-    error: courseError,
-  } = useQuery<CourseDetails>(CourseQuery, {
-    variables: { slug: clone },
+    loading,
+    error,
+    coursesData,
+    studyModulesData,
+    courseData,
+  } = useEditorCourses({
+    slug: clone,
   })
 
   useEffect(() => {
@@ -75,12 +56,8 @@ const NewCourse = () => {
     setClonedCourse({ ...stripId(courseData.course), slug: "" })
   }, [courseData])
 
-  if (studyModulesError || coursesError) {
-    return (
-      <ModifiableErrorMessage
-        errorMessage={JSON.stringify(studyModulesError || coursesError)}
-      />
-    )
+  if (error) {
+    return <ModifiableErrorMessage errorMessage={JSON.stringify(error)} />
   }
 
   return (
@@ -89,7 +66,7 @@ const NewCourse = () => {
         <H1NoBackground component="h1" variant="h1" align="center">
           {t("createCourse")}
         </H1NoBackground>
-        {studyModulesLoading || coursesLoading || courseLoading ? (
+        {loading ? (
           <FormSkeleton />
         ) : beta ? (
           <CourseEdit2
