+++ frontend/pages/_app.js
@@ -8,8 +8,8 @@ import { ThemeProvider } from "@material-ui/core/styles"
 import { ApolloProvider } from "@apollo/client"
 import Layout from "./_layout"
 import { isSignedIn, isAdmin } from "../lib/authentication"
-import LoginStateContext from "../contexes/LoginStateContext"
-import LanguageContext from "../contexes/LanguageContext"
+import LoginStateContext from "../contexts/LoginStateContext"
+import LanguageContext from "../contexts/LanguageContext"
 import withApolloClient from "../lib/with-apollo-client"
 import theme from "../src/theme"
 import { config as fontAwesomeConfig } from "@fortawesome/fontawesome-svg-core"
@@ -17,7 +17,7 @@ import "@fortawesome/fontawesome-svg-core/styles.css"
 import { CssBaseline } from "@material-ui/core"
 import PageTranslations from "/translations/pages"
 import { ConfirmProvider } from "material-ui-confirm"
-import AlertContext from "../contexes/AlertContext"
+import AlertContext from "../contexts/AlertContext"
 import getTranslator from "/translations"
 import { CacheProvider } from "@emotion/react"
 import createCache from "@emotion/cache"
